requirements.txt
pyotp
cryptography
requests
fastapi
uvicorn
pycryptodome
